+++ resources/views/emails/pin_login-plain.blade.php
@@ -3,9 +3,11 @@
 @section('content')
 @lang('ig-user::pin_login.intro')
 
-<x-ig-common::emails.button-plain :link="$loginUrl">{{ __('ig-user::pin_login.action') }}</x-ig-common::emails.button-plain>
+{{ $pin }}
 
-@lang('ig-user::pin_login.pin_label') {{ $pin }}
+@lang('ig-user::pin_login.login_page_label')
+
+{{ $loginUrl }}
 
 @lang('ig-user::pin_login.expires')
 @parent